pyproject.toml
[project]
name = "aoc"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  "numpy==2.2.0",
  "matplotlib==3.9.3",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "ruff>=0.8.2",
]

[project.scripts]
aoc = "aoc:both"
p1 = "aoc:part1"
p2 = "aoc:part2"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
